use parse-duration to sort time
