+++ data/new-list-single.json
@@ -389,7 +389,7 @@
             },
             {
                 "选手": "沐沐来了",
-                "成绩": "1'21''67",
+                "成绩": "1'19''36",
                 "日期": "2025/9/19"
             },
             {
