+++ data/new-list-total.json
@@ -1,18 +1,5 @@
 [
     {
-        "选手": "山月为关",
-        "三虎": "53''79",
-        "四僧": "1'26''75",
-        "四龙": "1'54''15",
-        "六将": "2'31''50",
-        "六虫": "2'14''44",
-        "万凶": "2'46''08",
-        "心猿": "1'21''98",
-        "梅山": "3'03''56",
-        "六根": "5'05''88",
-        "总成绩": "21'18''13"
-    },
-    {
         "选手": "馒头",
         "三虎": "54''52",
         "四僧": "1'28''78",
@@ -24,5 +11,18 @@
         "梅山": "3'23''94",
         "六根": "5'03''94",
         "总成绩": "21'40''78"
+    },
+    {
+        "选手": "山月为关",
+        "三虎": "53''79",
+        "四僧": "1'26''75",
+        "四龙": "1'54''15",
+        "六将": "2'31''50",
+        "六虫": "2'14''44",
+        "万凶": "2'46''08",
+        "心猿": "1'21''98",
+        "梅山": "3'03''56",
+        "六根": "5'05''88",
+        "总成绩": "21'18''13"
     }
 ]
