+++ package.json
@@ -13,10 +13,10 @@
     "doc": "docs"
   },
   "scripts": {
-    "dev": "node scripts/generate-list-data.cjs && vitepress dev docs",
-    "build": "node scripts/generate-list-data.cjs && vitepress build docs",
-    "preview": "node scripts/generate-list-data.cjs && vitepress build docs && vitepress preview docs",
-    "generate-list": "node scripts/generate-list-data.cjs"
+    "dev": "node scripts/generate-list-data.js && vitepress dev docs",
+    "build": "node scripts/generate-list-data.js && vitepress build docs",
+    "preview": "node scripts/generate-list-data.js && vitepress build docs && vitepress preview docs",
+    "generate-list": "node scripts/generate-list-data.js"
   },
   "dependencies": {
     "@vueuse/core": "^13.9.0"
@@ -28,6 +28,7 @@
     "gray-matter": "^4.0.3",
     "json2md": "^2.0.3",
     "markdown-it": "^14.1.0",
+    "parse-duration": "^2.1.4",
     "vitepress": "^2.0.0-alpha.6",
     "vitepress-plugin-group-icons": "^1.6.3",
     "vitepress-plugin-llms": "^1.7.3",
